+++ elements/src/components/pos-container-contents/pos-container-toolbar/shoelace.ts
@@ -1,0 +1,2 @@
+import '@shoelace-style/shoelace/dist/components/tooltip/tooltip.js';
+import '@shoelace-style/shoelace/dist/components/icon/icon.js';